fix shifted overlay on predict result for 1080p
The shifting happens on segmentation result, not on detection result.

The shape of output tensor for segmentation, which is the shape of mask,
is shown as 640(H)x1088(W), while the shape of input image is
1080(H)x1920(W).
So the ratios between 1080x1920 and 640x1088 are different, and the
difference leads to the shifted overlay.

When we apply the input ratio (W/H), which is approximately 1.77777, on the
shape of mask,
Height = 1088/1.77777 = 612
and we can assume there are pads on the top and bottom ends, 14 each.
So we eliminate the pads on both ends then scale up the mask to the size
of input image, which is 1080x1920.